SL-155/API-version-increase-fixes
