+++ controllers/front/successHosted.php
@@ -76,7 +76,8 @@ class SaferPayOfficialSuccessHostedModuleFrontController extends AbstractSaferPa
             if (
                 !$authResponseBody->getLiability()->getLiabilityShift() &&
                 in_array($order->payment, SaferPayConfig::SUPPORTED_3DS_PAYMENT_METHODS) &&
-                $paymentBehaviourWithout3DS === SaferPayConfig::PAYMENT_BEHAVIOR_WITHOUT_3D_CANCEL
+                $paymentBehaviourWithout3DS === SaferPayConfig::PAYMENT_BEHAVIOR_WITHOUT_3D_CANCEL ||
+                $authResponseBody->getTransaction()->getStatus() === 'CANCELED'
             ) {
                 $orderStatusService->cancel($order);
